// client/src/app/enums/route.ts
export enum RouteEnum {
    StartupError = 'startup-error',
    Login = 'login',
    Logout = 'logout',
    AccessRequest = 'access-request',
    Register = 'register',
    Home = 'home',
    Accounts = 'accounts',
    Transactions = 'transactions',
    Spending = 'spending',
    Investments = 'investments',
    Liabilities = 'liabilities',
    Admin = 'admin',
}